empty message

+++ src/main/java/com/jiker/keju/TaxPayment.java
@@ -34,7 +34,7 @@ public class TaxPayment {
         }
 
         if (distance <= 8) {
-            return (int) (BASE_PRICE + (distance - 2) * 0.8);
+            return BASE_PRICE + (distance - 2) * 0.8;
         }
         
         return (int) (distanceAmount(8) + (distance - 8) * 0.8 * 1.5);
